UPSTREAM: soc/amd/common: fix eSPI virtual wire polarity encoding
eSPI interrupts are active level high. The eSPI polarity register
in the chipset inverts incoming signals if the corresonding bit
is 0 in the register. Therefore, all active high (edge or level)
virtual wire interrupts need to ensure they are not inverted.

And really the sender of the interrupts should be conforming to the
the eSPI spec. As such inverting any signals should not be necessary,
but this register in the chipset allows for fixing up those misbehaviors.

BUG=b:157984427

Change-Id: I41d86e285a32d000364a602961946998fd0df304
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d3758540a951510ede21ebb76cbc196ae8ed0e68
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Change-Id: I7346bb0484506d96d7ab2e6d046ffa0571683a48
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43010
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2283066
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>