UPSTREAM: arch/x86/gdt: Work around assembler bug
The GDT loading did work fine on x86_64 a few months ago, but today it
only works in QEMU, but not on real hardware or KVM-enabled QEMU. This
might be related to toolchain changes.

Use 64bit GDT loading on x86_64 and force the assembler to generate a
64bit address load on the GDT. This will make sure no 32bit (signed)
displacement op is being generated, which points to the wrong address
in longmode.

Verified using readelf and made sure no R_X86_64_32S relocation symbol
is emitted. Disassembled the romstage ELF and made sure the GDT address
is 64bit in size.

Tested on QEMU and KVM-enabled QEMU: Doesn't crash any more on KVM.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iec858cdc0e34a89c3ad8d31c48189259e084d081
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 29ed4f56b5c41231081d2fccdf500dd307a3a8d3
Original-Signed-off-by: Patrick Rudolph <siro@das-labor.org>
Original-Change-Id: Ia824f90d9611e6e8db09bd62a05e6f990581f09a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43136
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2372234
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>