UPSTREAM: libpayload: allow nonblocking delay and more than one delay
Extend the local APIC timer delay so that it can be started,
and waited for, independently.

Add an EOI so that more than one APIC timer interrupt is possible.
Previous to this, because there was no EOI, the first timer
interrupt the CPU took was also the last it would take --
apic_delay would only work one time.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: a92a07d9c0a6079818385de78930a69459e4ee8a
Original-Change-Id: Ib11aeee5b7da81287166ac68fc327e7ae62d1b84
Original-Signed-off-by: Ronald G Minnich <rminnich@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43323
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: David Hendricks <david.hendricks@gmail.com>
Original-Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Change-Id: I288aeb80ea9da87206ef5d5e7d5a4c806f6c6ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2304836
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>