UPSTREAM: ec/google/chromeec: Fix Coverity Scan error (BAD_SHIFT)
A recent Coverity scan found an issue with the way the
EC_HOST_EVENT_MASK macro was being used. It was being passed values
between 0 and 63, but since it is doing basically (1ULL << (value - 1)),
this caused a shift of -1 when `i` is 0 and also doesn't reach the 63rd
bit of the mask. This is fixed by incrementing the start and end
conditions of the loop by 1, so the event mask ranges from bits 0 to 63,
instead of -1 to 62.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 7777e1c30b5c9d47388627c080ff86a4e043ec7e
Original-Found-by: Coverity CID 1430218
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I6a7cfa64545f3d313de24407f0a91b48368f2a8a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43460
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Change-Id: I59b1ef42fce861218aa963334986b9401d8fc163
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2315554
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>