UPSTREAM: mb/google/zork: Drop redundant romstage GPIO table
Now that the power and reset GPIO configuration for non-PCIe devices is
dropped from romstage GPIO table, the tables for pre-v3 and v3 version of
schematics are exactly same. So, this change drops the duplicate table and
also removes the check for v3 schematics when configuring the pads in
romstage.

BUG=b:154351731

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 6a5c77cc846384dbc10c3546f70245025787ef08
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I67ca9f587c3f47912393ebaf38badcc9d76cc393
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43474
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Change-Id: I995c94e351b6a6333b6730083f0f1107b79eb8e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2304832
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>