UPSTREAM: google/trogdor: Remove GPIO_AP_SUSPEND_L
This pin has been removed from recent revisions (or at least moved to a
pin that's not controlled like a normal GPIO). It has also never been
used on older ones (I think?). The same GPIO is now used as a SKU ID
pin, so make sure we're not incorrectly configuring it as an output
(although sku_id() overrides that later anyway, but it's still
incorrect).

BUG=b:160754995

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: aa4c1d2b5b3dd210e09239fb28e5cd0a8bd6b0fc
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I6e0d56e230d0bef172d7b78cc48263e9b6f059de
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43471
Original-Reviewed-by: Douglas Anderson <dianders@chromium.org>
Original-Reviewed-by: Philip Chen <philipchen@google.com>
Original-Reviewed-by: Bob Moragues <moragues@google.com>
Original-Reviewed-by: EricR Lai <ericr_lai@compal.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I264547c0694d452a3f8925aec6f3a43a24a490c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2304181
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>