UPSTREAM: dptf: Fix scope of TCPU device
In the initial DPTF refactor, the scope of the TCPU device was
incorrectly set as \_SB, instead of \_SB.PCI0. However, because of the
way that the acpi_inject_dsdt() callback currently works (it injects
contents before the dsdt.aml file), the Scope where the TCPU
device lives (\_SB.PCI0) doesn't exist yet. Therefore, to avoid playing
games with *when* things are defined in the DSDT, switch to defining all
of the DPTF devices in the SSDT.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 5212ece6cf27fa91ffd07d97d4fbe171129cf1fa
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: Ia4922b4dc6544d79d44d39e6ad18c6ab9fee0fd7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43529
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: Icc68c6f4b9acd1eaeba6d3685ddd5a56d284caaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2311175
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>