UPSTREAM: lib: Add ASan support to romstage on x86 arch
This patch adds ASan support to romstage on x86 architecture.
A Kconfig option is added to enable ASan in romstage. Compiler
flags are updated. A memory space representing the shadow region
is reserved in linker section. And a function call to asan_init()
is added to initialize shadow region when romstage loads.

BUG=none
BRANCH=none
TEST=none

Change-Id: Idd295bf8bb1249853e78800a4e7ba05e071b4a6c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a6ebe08333dcc6c263b6bb8936d4883663c63798
Original-Change-Id: I67ebfb5e8d602e865b1f5c874860861ae4e54381
Original-Signed-off-by: Harshit Sharma <harshitsharmajs@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43604
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2372221
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>