UPSTREAM: soc/amd/picasso/Makefile.inc: force an error if PSPBTLDR_FILE is not set
Currently, if PSPBTLDR_FILE is empty, the md5sum will hang forever on
stdin, leading to the appearance of a hung script. This is
confusing.

There's no option to md5sum to say "you must use this file", so instead,
use dd with if to ensure we at least get an error if the file is not found.

Not optimal, but better than what we have now.

BUG=none
BRANCH=none
TEST=none

Change-Id: I68f985604d38f0306d1451743b6e2cdd8a44fa26
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d6326a0faabc3cc3d688317653cf35cf4815ed13
Original-Change-Id: Ia13035bc592bdf2a515dfd2e052ae9135e218612
Original-Signed-off-by: Ronald G Minnich <rminnich@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43677
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jeremy Soller <jeremy@system76.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2324917
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>