UPSTREAM: security/vboot/Makefile.inc: Update regions-for-file function
This patch updates regions-for-file function in the
security/vboot/Makefile.inc to support adding a CBFS file into
required FMAP REGIONs in a flexible manner. The file that needs to be
added to specific REGIONs, those regions list should be specified in the
regions-for-file-{CBFS_FILE_TO_BE_ADDED} variable.

For example, if a file foo.bin needs to be added in FW_MAIN_B and COREBOOT,
then below code needs to be added in a Makefile.inc.
        regions-for-file-foo := FW_MAIN_B,COREBOOT
        cbfs-file-y := foo
        foo-file := foo.bin
        foo-type := raw

TEST=Verified on hatch

Change-Id: I74a94a4df535e2540fe3ed4d5811555c8505c0b1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 77025b3f56ab599db70e923650eb6e42d84f267f
Original-Signed-off-by: Sridhar Siricilla <sridhar.siricilla@intel.com>
Original-Change-Id: I1f5c22b3d9558ee3c5daa2781a115964f8d2d83b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43766
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2359434
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>