UPSTREAM: src: Change BOOL CONFIG_ to CONFIG() in comments & strings
The Kconfig lint tool checks for cases of the code using BOOL type
Kconfig options directly instead of with CONFIG() and will print out
warnings about it.  It gets confused by these references in comments
and strings.  To fix it so that it can find the real issues, just
update these as we would with real issues.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0b236f15a3f9536bac5811b9ae41d9cd13551fe1
Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: f48acbda7be7074938c06db8ad37705f850661ee
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I5c37f0ee103721c97483d07a368c0b813e3f25c0
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43824
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2321680
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>