UPSTREAM: drivers/intel/dptf: Remove prompts from DPTF config options
The prompts for the DPTF Kconfig options were not necessary, they should
be selected based on what DPTF implementation is being used, ASL files
or generated at runtime. It's not really meant to be fiddled with at
build-time. Also rewrite the help text for the _HID selection, to try
and make it more clear when to use y or n.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia264e5ee9d373221e6eb797e780e9ca219cd47f1
Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 0aabd07c95c3a940832d2818bacb47843388ad85
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I6edcabd28426916d9586d501b95b510dfc163fc1
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/43830
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2321695
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>