UPSTREAM: soc/amd/picasso: If using VBOOT, skip the APOB_NV region for RO
When booting from the RO region of a VBOOT enabled ROM, there shouldn't
be a reliance on anything outside of the RO section.  This includes the
APOB_NV region (similar to the MRC cache region).  By skipping the
region when setting up the BIOS Directory table, the PSP won't try to
use the region when booting.

The APOB_NV region is still used for the VBOOT RW sections.

BUG=b:158363448
TEST=Build RO with no APOB_NV region.  Dump the BDT and verify that
it's not in RO, but is in RW_A & RW_B.  Boot into recovery.

Change-Id: Icb3a652d9244373bd5ecf63a41a15f3b5f8ebf26
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e2d0ba0acbc6b2b6c4d9f350f8e67b2287054f79
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: I13c35ba8a2331492744d2acf257db15e4a53102a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44046
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2372240
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>