UPSTREAM: assert.h: Try to evaluate assertions at compile time
Many places in coreboot seem to like to do things like

 assert(CONFIG(SOME_KCONFIG));

This is somewhat suboptimal since assert() is a runtime check, so you
don't see that this fails until someone actually tries to boot it even
though the compiler is totally aware of it already. We already have the
dead_code() macro to do this better:

 if (CONFIG(SOME_KCONFIG))
   dead_code();

Rather than fixing all these and trying to carefully educate people
about which type of check is more appropriate in what situation, we can
just employ the magic of __builtin_constant_p() to automatically make
the former statement behave like the latter.

BUG=none
BRANCH=none
TEST=none

Change-Id: If07a5352f798e0db8395fb65ea31c11cc6e3bf3d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2aedc9776a3c2ef99c8ca4f1ed836c3f7f6821d6
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I06691b732598eb2a847a17167a1cb92149710916
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44044
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2334915
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>