UPSTREAM: mb/google/volteer: Pull up GPP_D16 instead of driving it
The latest realtek RTS5261 SD daughterboard exposes the PRSNT# pin to
GPP_D16 but there is a RTS5261 requirement to pull up this pin and not
drive it at power on. We can meet this requirement without breaking
other boards by changing GPP_D16 to be a no-connect with an internal
pull up. Other boards use this signal as an enable input, so changing
this to pull up is OK.

BUG=b:162722965
TEST=Verified RTS5261 and GL9755 daughterboards enumerate on PCI and
        can read SD cards.

Change-Id: I0890b08f9f8b00e3b27bd8b6b0d096afdbd4bdfa
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>
Original-Commit-Id: 173493784dd07ac1cbf055a0b26c55f9bd1f5a28
Original-Change-Id: I096d76ec12b7c3afaf02e621fd301b6704913d5d
Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44116
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2347849