UPSTREAM: gpio: Pull down HiZ pins after reading tristate GPIO strapping
People who know a lot more about electrons and stuff than I do tell me
that leaving a HiZ pin floating without a pull resistor may waste power.
So if we find a pin to be HiZ when reading tristate strapping GPIOs, we
should make sure the internal pull-down is enabled when we're done with
it. (For pins that are externally pulled high or low, we should continue
to leave the internal pull disabled instead.)

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 7450790558f3b1f85e9a11df252dd27f1047bfeb
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I1669823c8a7faab536e0441cb4c6cfeb9f696189
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44253
Original-Reviewed-by: Douglas Anderson <dianders@chromium.org>
Original-Reviewed-by: Alexandru Stan <amstan@google.com>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: Ic4845e970558383f82b0a2da4cc74d3d5aea1b4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2343928
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>